% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{cut.off.single}
\alias{cut.off.single}
\title{Single Sample Cut Off}
\usage{
\method{cut}{off.single}(mag.output)
}
\arguments{
\item{mag.output}{This is the output of the mag.single function}
}
\value{
A list of R files. The main output is final.data which has the samples and the cluster assingment in a dataframe.
}
\description{
This function is used to find the best cut off and the cluster assignment on single sample.
}
