% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{find.cut.off.multiple}
\alias{find.cut.off.multiple}
\title{Find cut off- multiple sample.}
\usage{
find.cut.off.multiple(mag.var, shape, depth)
}
\description{
Finding the cut off of the multiple sample clustering.
}
