% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{find.cut.off}
\alias{find.cut.off}
\title{Single Sample Cut Off}
\usage{
find.cut.off(mag.output, shape, depth)
}
\arguments{
\item{mag.output}{This is the output of the mag.single function}

\item{shape}{Shape corresponds to the variance of the germline sample variance. In most experiments we take shape to be around 8 or 10.}

\item{depth}{Depth is the sequence coverage of the experiment. By default the algorithm uses depth of 200.}
}
\value{
A list of R files. The main output is final.data which has the samples and the cluster assingment in a dataframe.
}
\description{
This function is used to find the best cut off and the cluster assignment on single sample.
}
