% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{mag.multiple.cut}
\alias{mag.multiple.cut}
\title{MAG Multiple Sample Clustering}
\usage{
mag.multiple.cut(arg.data, x.el.cut = data.frame(), sampleNum = 1)
}
\arguments{
\item{arg.data}{Preproccessed dataframe of the data.}

\item{x.el.cut}{Is the x.el.red from the mag.reduce.graph(). This can also be empty.}

\item{sampleNum}{This parameter corresponds to the column of the dataframe that the main sample is stored in.}
}
\value{
A list of resutls. A list of dataframes of cluster assignments at each step of hierarchical step. The first element of x.el.list the input x.el.cut.
}
\description{
This function performs clustering on multiple samples with the focus of one main sample.
}
