% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{mag.multiple}
\alias{mag.multiple}
\title{MAG Multiple Sample Clustering}
\usage{
mag.multiple(prep.data, x.el.cut = data.frame())
}
\arguments{
\item{prep.data}{Preproccessed dataframe of the data.}

\item{x.el.cut}{Is the x.el.red from the mag.reduce.graph(). This can also be empty.}
}
\description{
This function performs clustering on multiple samples with the focus of one main sample.
}
