% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{mag.multiple.run}
\alias{mag.multiple.run}
\title{MAG run}
\usage{
mag.multiple.run(input.data, reduce = T, threshold = 0.03, fold = F)
}
\description{
The main function of the package. After performing the preprocessing this function performs all steps of clustereing and finding the best structure and grouping of mutations.
}
