% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{mag.prepdata}
\alias{mag.prepdata}
\title{MAG data preparation}
\usage{
mag.prepdata(arg.data)
}
\arguments{
\item{arg.data}{Dataframe of the samples. Each sample should be in a separate column.}
}
\value{
Returns a dataframe with processed data and ID.
}
\description{
This function preprocesses the input data for the main algorithm.
}
