% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{mag.reduce.graph}
\alias{mag.reduce.graph}
\title{MAG REDUCE STEP}
\usage{
mag.reduce.graph(arg.data, x.el = data.frame(), sampleNum = 1, n = -100)
}
\arguments{
\item{arg.data}{Preproccessed dataframe of the data.}

\item{x.el}{A dataframe of samples that want to be forced together. Each row coresponds to a cluster and each column represents a data point (mutation). This is a 0-1 matrix.
Sum of each column should be 1 and sum of each row should be number of mutations in that cluster. At step 0, where each mutation is a cluster, this dataframe is a m*m matrix with 1s on the diagonal. The default is an empty dataframe but the function populates this dataframe.}

\item{sampleNum}{This parameter corresponds to the column of the dataframe that the sample is stored in.}

\item{n}{This variable is used internaly. Do not change it.}
}
\value{
A list of files. x.el.red is the main output of the function. It is the update dataframe of clusters and mutations after performing the initial merging.
}
\description{
This function is used to create a pre clustering merging of the data. This function is used to merge data points that are close to each other using a graph approach.
}
