% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{mag.run}
\alias{mag.run}
\title{MAG run}
\usage{
mag.run(data, name = "", sampleFreqs = NA, sampleNum = 1, shape = 10,
  depth = 200)
}
\description{
The main function of the package. After performing the preprocessing this function performs all steps of clustereing and finding the best structure and grouping of mutations.
}
