% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{mag.single}
\alias{mag.single}
\title{MAG single}
\usage{
mag.single(arg.data, sampleNum = 1)
}
\arguments{
\item{arg.data}{Preproccessed dataframe of the data.}

\item{sampleNum}{This parameter corresponds to the column of the dataframe that the sample is stored in.}
}
\value{
List of data files. List of matrixs of each step of the hierarchical clustering. Sorted sample. A dataframe of the variance and mean frequency of samples at each clustering step.
}
\description{
This function is used to perform clustering on only one sample.
}
