% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{mag.single.run}
\alias{mag.single.run}
\title{MAG run}
\usage{
mag.single.run(input.data, fold = F)
}
\description{
The main function of the package. After performing the preprocessing this function performs all steps of clustereing and finding the best structure and grouping of mutations.
}
