% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{mag.prepdata}
\alias{mag.prepdata}
\title{MAG data preparation}
\usage{
mag.prepdata(arg.data)
}
\arguments{
\item{arg.data}{Dataframe of the samples. Each column pairs correspond to reference and alternate counts. Column 1 and 2 for sample 1, column 3 and 4 for sample 2 and ...}
}
\value{
Returns a dataframe with processed data and ID.
}
\description{
This function preprocesses the input data for the main algorithm.
}
