% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGOS.R
\name{mag.single}
\alias{mag.single}
\title{MAG single}
\usage{
mag.single(prep.data)
}
\arguments{
\item{prep.data}{Preproccessed dataframe of the data.}
}
\value{
List of data files. List of matrixs of each step of the hierarchical clustering. Sorted sample. A dataframe of the variance and mean frequency of samples at each clustering step.
}
\description{
This function is used to perform clustering on only one sample.
}
